requirements.txt
aiogram==2.20
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
djangorestframework==3.13.1
frozenlist==1.3.0
geographiclib==1.52
geopy==2.2.0
idna==3.3
multidict==6.0.2
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
telebot==0.0.4
urllib3==1.26.9
yarl==1.7.2

folium~=0.12.1.post1
geocoder~=1.38.1
gmaps~=0.9.0